requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
dnspython==2.4.2 